+++ flow/test/unit/driver/chunk.cpp
@@ -13,6 +13,7 @@
 #include <gtest/gtest.h>
 
 // Flow
+#include <flow/captor/nolock.h>
 #include <flow/driver/chunk.h>
 
 using namespace flow;
