+++ flow/test/unit/follower/before.cpp
@@ -12,6 +12,7 @@
 #include <gtest/gtest.h>
 
 // Flow
+#include <flow/captor/nolock.h>
 #include <flow/follower/before.h>
 
 using namespace flow;
