+++ flow/test/unit/follower/latched.cpp
@@ -13,6 +13,7 @@
 #include <gtest/gtest.h>
 
 // Flow
+#include <flow/captor/nolock.h>
 #include <flow/follower/latched.h>
 
 using namespace flow;
