+++ flow/test/unit/follower/ranged.cpp
@@ -14,6 +14,7 @@
 #include <gtest/gtest.h>
 
 // Flow
+#include <flow/captor/nolock.h>
 #include <flow/follower/ranged.h>
 
 using namespace flow;
